+++ src/crypto/rsa.rs
@@ -1,5 +1,4 @@
 use ring::{rand, signature};
-use simple_asn1::BigUint;
 
 use crate::algorithms::Algorithm;
 use crate::errors::{ErrorKind, Result};
@@ -37,30 +36,27 @@ pub(crate) fn alg_to_rsa_signing(alg: Algorithm) -> &'static dyn signature::RsaE
 pub(crate) fn sign(
     alg: &'static dyn signature::RsaEncoding,
     key: &[u8],
-    message: &str,
+    message: &[u8],
 ) -> Result<String> {
-    let key_pair = signature::RsaKeyPair::from_der(key).map_err(|_| ErrorKind::InvalidRsaKey)?;
+    let key_pair = signature::RsaKeyPair::from_der(key)
+        .map_err(|e| ErrorKind::InvalidRsaKey(e.description_()))?;
 
     let mut signature = vec![0; key_pair.public_modulus_len()];
     let rng = rand::SystemRandom::new();
-    key_pair
-        .sign(alg, &rng, message.as_bytes(), &mut signature)
-        .map_err(|_| ErrorKind::InvalidRsaKey)?;
+    key_pair.sign(alg, &rng, message, &mut signature).map_err(|_| ErrorKind::RsaFailedSigning)?;
 
-    Ok(b64_encode(&signature))
+    Ok(b64_encode(signature))
 }
 
 /// Checks that a signature is valid based on the (n, e) RSA pubkey components
 pub(crate) fn verify_from_components(
     alg: &'static signature::RsaParameters,
     signature: &str,
-    message: &str,
-    components: (&str, &str),
+    message: &[u8],
+    components: (&[u8], &[u8]),
 ) -> Result<bool> {
     let signature_bytes = b64_decode(signature)?;
-    let n = BigUint::from_bytes_be(&b64_decode(components.0)?).to_bytes_be();
-    let e = BigUint::from_bytes_be(&b64_decode(components.1)?).to_bytes_be();
-    let pubkey = signature::RsaPublicKeyComponents { n, e };
-    let res = pubkey.verify(alg, message.as_ref(), &signature_bytes);
+    let pubkey = signature::RsaPublicKeyComponents { n: components.0, e: components.1 };
+    let res = pubkey.verify(alg, message, &signature_bytes);
     Ok(res.is_ok())
 }