+++ src/encoding.rs
@@ -9,6 +9,7 @@ use crate::serialization::b64_encode_part;
 
 /// A key to encode a JWT with. Can be a secret, a PEM-encoded key or a DER-encoded key.
 /// This key can be re-used so make sure you only initialize it once if you can for better performance
+/// Currently only PKCS8 formatted EC keys are supported, ensure this format is selected when generating the key pair.
 #[derive(Debug, Clone, PartialEq)]
 pub struct EncodingKey {
     pub(crate) family: AlgorithmFamily,
@@ -60,6 +61,14 @@ impl EncodingKey {
         Ok(EncodingKey { family: AlgorithmFamily::Ec, content: content.to_vec() })
     }
 
+    /// If you are loading a EdDSA key from a .pem file
+    /// This errors if the key is not a valid private Ed key
+    pub fn from_ed_pem(key: &[u8]) -> Result<Self> {
+        let pem_key = PemEncodedKey::new(key)?;
+        let content = pem_key.as_ed_private_key()?;
+        Ok(EncodingKey { family: AlgorithmFamily::Ed, content: content.to_vec() })
+    }
+
     /// If you know what you're doing and have the DER-encoded key, for RSA only
     pub fn from_rsa_der(der: &[u8]) -> Self {
         EncodingKey { family: AlgorithmFamily::Rsa, content: der.to_vec() }
@@ -68,6 +77,11 @@ impl EncodingKey {
     /// If you know what you're doing and have the DER-encoded key, for ECDSA
     pub fn from_ec_der(der: &[u8]) -> Self {
         EncodingKey { family: AlgorithmFamily::Ec, content: der.to_vec() }
+    }
+
+    /// If you know what you're doing and have the DER-encoded key, for EdDSA
+    pub fn from_ed_der(der: &[u8]) -> Self {
+        EncodingKey { family: AlgorithmFamily::Ed, content: der.to_vec() }
     }
 
     pub(crate) fn inner(&self) -> &[u8] {
@@ -101,10 +115,10 @@ pub fn encode<T: Serialize>(header: &Header, claims: &T, key: &EncodingKey) -> R
     if key.family != header.alg.family() {
         return Err(new_error(ErrorKind::InvalidAlgorithm));
     }
-    let encoded_header = b64_encode_part(&header)?;
-    let encoded_claims = b64_encode_part(&claims)?;
-    let message = [encoded_header.as_ref(), encoded_claims.as_ref()].join(".");
-    let signature = crypto::sign(&*message, key, header.alg)?;
+    let encoded_header = b64_encode_part(header)?;
+    let encoded_claims = b64_encode_part(claims)?;
+    let message = [encoded_header, encoded_claims].join(".");
+    let signature = crypto::sign(message.as_bytes(), key, header.alg)?;
 
     Ok([message, signature].join("."))
 }