+++ src/lib.rs
@@ -16,12 +16,10 @@ mod serialization;
 mod validation;
 
 pub use algorithms::Algorithm;
-#[allow(deprecated)]
-pub use decoding::dangerous_unsafe_decode;
 pub use decoding::{
     dangerous_insecure_decode, dangerous_insecure_decode_with_validation, decode, decode_header,
-    DecodingKey, TokenData,
+    verify_signature, DecodingKey, TokenData,
 };
 pub use encoding::{encode, EncodingKey};
 pub use header::Header;
-pub use validation::Validation;
+pub use validation::{get_current_timestamp, Validation};