+++ tests/rsa/mod.rs
@@ -28,10 +28,14 @@ fn round_trip_sign_verification_pem_pkcs1() {
 
     for &alg in RSA_ALGORITHMS {
         let encrypted =
-            sign("hello world", &EncodingKey::from_rsa_pem(privkey_pem).unwrap(), alg).unwrap();
-        let is_valid =
-            verify(&encrypted, "hello world", &DecodingKey::from_rsa_pem(pubkey_pem).unwrap(), alg)
-                .unwrap();
+            sign(b"hello world", &EncodingKey::from_rsa_pem(privkey_pem).unwrap(), alg).unwrap();
+        let is_valid = verify(
+            &encrypted,
+            b"hello world",
+            &DecodingKey::from_rsa_pem(pubkey_pem).unwrap(),
+            alg,
+        )
+        .unwrap();
         assert!(is_valid);
     }
 }
@@ -43,10 +47,14 @@ fn round_trip_sign_verification_pem_pkcs8() {
 
     for &alg in RSA_ALGORITHMS {
         let encrypted =
-            sign("hello world", &EncodingKey::from_rsa_pem(privkey_pem).unwrap(), alg).unwrap();
-        let is_valid =
-            verify(&encrypted, "hello world", &DecodingKey::from_rsa_pem(pubkey_pem).unwrap(), alg)
-                .unwrap();
+            sign(b"hello world", &EncodingKey::from_rsa_pem(privkey_pem).unwrap(), alg).unwrap();
+        let is_valid = verify(
+            &encrypted,
+            b"hello world",
+            &DecodingKey::from_rsa_pem(pubkey_pem).unwrap(),
+            alg,
+        )
+        .unwrap();
         assert!(is_valid);
     }
 }
@@ -57,9 +65,10 @@ fn round_trip_sign_verification_der() {
     let pubkey_der = include_bytes!("public_rsa_key.der");
 
     for &alg in RSA_ALGORITHMS {
-        let encrypted = sign("hello world", &EncodingKey::from_rsa_der(privkey_der), alg).unwrap();
+        let encrypted = sign(b"hello world", &EncodingKey::from_rsa_der(privkey_der), alg).unwrap();
         let is_valid =
-            verify(&encrypted, "hello world", &DecodingKey::from_rsa_der(pubkey_der), alg).unwrap();
+            verify(&encrypted, b"hello world", &DecodingKey::from_rsa_der(pubkey_der), alg)
+                .unwrap();
         assert!(is_valid);
     }
 }
@@ -108,7 +117,7 @@ fn rsa_modulus_exponent() {
     .unwrap();
     let res = decode::<Claims>(
         &encrypted,
-        &DecodingKey::from_rsa_components(n, e),
+        &DecodingKey::from_rsa_components(n, e).unwrap(),
         &Validation::new(Algorithm::RS256),
     );
     assert!(res.is_ok());